+++ applications/payment-service/src/main/java/com/trading/payments/entity/Transaction.java
@@ -21,7 +21,7 @@ public class Transaction {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;  // Changed from Long to Integer to match SERIAL
     
     @Column(name = "transaction_id", unique = true, nullable = false)
     private String transactionId;
